requirements.txt
Django>=5.2,<5.3
django-autoslug
django-embed-video
djangorestframework
djangorestframework_simplejwt
drf_yasg
django-cors-headers
celery[redis]
psycopg
asyncpg
Pillow
easy-thumbnails==2.10
python-dotenv
python-dateutil
requests
user-agents
moviepy
Faker
gunicorn
uvicorn